+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stoker-platform/web-app
 
+## 0.5.67
+
+### Patch Changes
+
+- feat: add rich text editor toolbar features
+
 ## 0.5.66
 
 ### Patch Changes